refactor: wizard step updates
